+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.7.4] - 3 May, 2024
-- Cross-site scripting (XSS) bugfix.
+## [1.7.5] - 28 May, 2024
+- Fixes a issue Banner and carousel templates regarding background color.
+
+## [1.7.4] - 17 May, 2024
+- Fixes a issue in path for installation of Service worker for shopify.
 
 ## [1.7.3] - 3 Apr, 2024
 - Fixes a issue in release pipeline of Service worker.